pyproject.toml
[project]
name = "test-blaek-ble"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "asgiref>=3.8.1",
    "bleak>=0.22.3",
    "flask[async]>=3.0.3",
    "hypercorn>=0.17.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
